// control-pnale/src/app/shard/models/IdataBouquet.ts
export interface IBou {
  playlistRequired: number;
  dailyWorkingHours: Date;
  numberOfGrids: number;
  monthlyInternetConsumption: number;
  daysOff: number;
  dailySubscriberSum: number;
  collectorOfDailyHours: number;
  price: number;
  playlist?: string;
  title: string;
  date?: Date;
  numberOfDays: number;
  keyword: string;
  des: string;
}